wifi: Make WifiPhyState a scoped enum within ns3 namespace

+++ src/energy/examples/basic-energy-model-test.cc
@@ -165,7 +165,10 @@ BasicEnergyUpdateTest::StateSwitchTest(WifiPhyState state)
      */
 
     // schedule change of state
-    Simulator::Schedule(Seconds(m_timeS), &WifiRadioEnergyModel::ChangeState, devModel, state);
+    Simulator::Schedule(Seconds(m_timeS),
+                        &WifiRadioEnergyModel::ChangeState,
+                        devModel,
+                        static_cast<int>(state));
 
     // Calculate remaining energy at simulation stop time
     Simulator::Schedule(Seconds(m_timeS * 2), &BasicEnergySource::UpdateEnergySource, source);
